add retract action
