+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/ClawSlide.java
@@ -13,10 +13,11 @@ public class ClawSlide {
     private static final double LIFT_POWER = 1.0;
     private static final int ROTATE_MAX_POSITION = 900;
     private static final double ROTATE_ANGLE_RATIO = 90.0 / 1000;
-    private static final int LIFT_MAX_POSITION = 2750;
+    private static final int LIFT_MAX_POSITION = 2900;
     private static final int LIFT_MAX_POSITION_HORIZON = 2150;
 
     private final Action PUT_DOWN_AND_EXTEND_ACTION;
+    private final Action RETRACT_AND_PULL_UP_ACTION;
 
     public final MotorPair slideRotate, slideLift;
     public final Claw claw;
@@ -45,9 +46,14 @@ public class ClawSlide {
         this.PUT_DOWN_AND_EXTEND_ACTION = new Action(
             new TimedAction(1.0, () -> ClawSlide.this.claw.setRotate(0)),
             new MotorPairAction(this.slideLift, 0),
-            new MotorPairAction(this.slideRotate, 1000),
+            new MotorPairAction(this.slideRotate, 850),
             new MotorPairAction(this.slideLift, LIFT_MAX_POSITION_HORIZON),
             new TimedAction(1.0, () -> ClawSlide.this.claw.setRotate(90)));
+
+        this.RETRACT_AND_PULL_UP_ACTION = new Action(
+            new TimedAction(1.0, () -> ClawSlide.this.claw.setRotate(90)),
+            new MotorPairAction(this.slideLift, 0),
+            new MotorPairAction(this.slideRotate, 0));
 
         this.slideRotate.resetPosition();
         this.slideLift.resetPosition();
@@ -66,12 +72,20 @@ public class ClawSlide {
         this.action = null;
     }
 
-    public void putDownAndExtend() {
+    private void setAction(Action action) {
         if (this.inAction()) {
             this.cancelAction();
         }
-        this.action = this.PUT_DOWN_AND_EXTEND_ACTION;
+        this.action = action;
         this.action.reset();
+    }
+
+    public void putDownAndExtend() {
+        this.setAction(this.PUT_DOWN_AND_EXTEND_ACTION);
+    }
+
+    public void retractAndPullUp() {
+        this.setAction(this.PUT_DOWN_AND_EXTEND_ACTION);
     }
 
     public void update() {
